Move test_util and fix header_generator_test
Moves test_util into compiler/util/ so it can be used by the backend
while minimizing importing frontend files. Additionally fixes the
following in header_generator_test:

 - Import test_util (tests failed due to missing import).
 - Change attribute_checker.normalize_and_verify to
   header_generator.generate_header.
 - Remove unused imports.

This change also adds a (blank) __init__.py to compiler/ and
compiler/util/ so that they are both importable by unittest. This means
that running `python -m unittest discover -p '*_test.py' from the main
emboss repository directory should run all python tests in the compiler.
