+++ compiler/front_end/glue_test.py
@@ -18,10 +18,10 @@ import pkgutil
 import unittest
 
 from compiler.front_end import glue
-from compiler.front_end import test_util
 from compiler.util import error
 from compiler.util import ir_pb2
 from compiler.util import parser_types
+from compiler.util import test_util
 
 _location = parser_types.make_location
